move direction to wlr_output_layout.h
