--- include/wlr/util/direction.h
@@ -1,11 +1,0 @@
-#ifndef WLR_UTIL_DIRECTION_H
-#define WLR_UTIL_DIRECTION_H
-
-enum wlr_direction {
-	WLR_DIRECTION_UP = 0,
-	WLR_DIRECTION_DOWN = 1,
-	WLR_DIRECTION_LEFT = 2,
-	WLR_DIRECTION_RIGHT = 4,
-};
-
-#endif
